CI: execute UT unit tests
Runs all registered UT tests. Uses an emscripten enabled ut fork for now
(draft PR pending). The tests are run through ctest and automatically
Use nodejs to run the tests in the emscripten build. This uses the
system nodejs instead of the one provided by emsdk, because the emsdk
was not able to run the tests
Disables asan depending on the compiler in the CMake because there are
unresolved issues reported.